+++ tests/testthat/test-vars-select.R
@@ -68,3 +68,8 @@ test_that("unknown variables errors are ignored if `.strict` is FALSE", {
   expect_identical(vars_select(letters, a, `_foo`, .strict = FALSE), c(a = "a"))
   expect_identical(vars_select(letters, c(a, `_foo`, c), .strict = FALSE), c(a = "a", c = "c"))
 })
+
+test_that("`:` handles strings", {
+  expect_identical(vars_select(letters, "b":"d"), vars_select(letters, b:d))
+  expect_error(vars_select(letters, "b":"Z"), "Unknown column `Z`")
+})